feat: add folder actionCreators and move there authActionCreators
